Create new cluster desc if path to cluster desc is stale
* [run_ci] Check if cluster_desc path exists

(cherry picked from commit 0c85eb6ff3b97e21336dfc6fdfbef9a14497f16b)

+++ loader/tt_cluster.cpp
@@ -96,7 +96,7 @@ std::string tt_cluster::get_cluster_desc_path(const std::string& root) {
     }
     if (!fs::exists(custom_cluster_desc_path) or !cluster_desc_path.size()) {
         // Either the cluster descriptor has not been generated or not been populated in the dir specified by root.
-        if (cluster_desc_path.size()) {
+        if (cluster_desc_path.size() && fs::exists(cluster_desc_path)) {
             // Cluster desc in root does not exist but has been previously generated and stored in cluster_desc_path.
             // Copy cluster desc to root.
             fs::copy(cluster_desc_path, custom_cluster_desc_path);
